Added alembic, added initial migration for weather_notifications table

+++ bot/models/weather_notifications.py
@@ -1,9 +1,9 @@
 from datetime import time
 from enum import Enum
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from sqlalchemy import Integer, Float, Time, Enum as SqlAlchemyEnum
+from sqlalchemy import Integer, BigInteger, Float, Time, Enum as SqlAlchemyEnum
 
-from base import Base
+from .base import Base
 
 
 class NotificationsType(Enum):
@@ -15,7 +15,7 @@ class WeatherNotifications(Base):
     __tablename__ = "weather_notifications"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    chat_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    chat_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     notifications_type: Mapped[NotificationsType] = mapped_column(SqlAlchemyEnum(NotificationsType))
     latitude: Mapped[float] = mapped_column(Float, nullable=False)
     longitude: Mapped[float] = mapped_column(Float, nullable=False)
